Move options parsing to own module
