+++ cardano-node-chairman/cardano-node-chairman.cabal
@@ -55,6 +55,7 @@ executable cardano-node-chairman
   hs-source-dirs:      app
   main-is:             cardano-node-chairman.hs
   other-modules:       Cardano.Chairman
+                       Cardano.Chairman.Options
                        Paths_cardano_node_chairman
   default-language:    Haskell2010
   ghc-options:         -threaded
